+++ src/data/codec/pack/wrap/mod.rs
@@ -3,14 +3,14 @@
 //! Structured wrappers and chunked containers.
 //
 
-// mod riff;
+mod riff; // Resource Interchange File Format
 // mod png_chunk;
 
 crate::structural_mods! { // _mods
     _mods {
-        // pub use super::{
-        //     riff::_all::*,
-        //     png_chunk::*,
-        // };
+        pub use super::{
+            riff::_all::*,
+            // png_chunk::*,
+        };
     }
 }